Remove hardcoded versions from pip requirements files
If we do not freeze the version we will automatically get the latest and
greatest of linting tools etc. We would need to update zivid-python
whenever any linter tools get a new test we cannot pass, but I think
this is a worthwhile trade.

The newer pip modules required some changes:
    - pytest-unmarked was replaced with calling pytest with -m
    - added Raises section in docstrings
    - ignoring pytest import-outside-toplevel
    - change error type to TypeError when initializing camera, etc