Add Github Actions workflow to assist release deployment
This commit adds a new workflow called deploy.yml which creates
deployable artifacts for PyPI (or direct installation with pip). The
workflow is triggered on push to master branch (which includes any time
a PR towards master is merged). This choice was made because it is
convenient that up-to-date artifacts always are available, even if there
is no impending release.

The workflow does not go as far as actually upload to PyPI. The purpose
of the workflow is to provide, as an artifact, a zip-file which contains
the full set of tested and validated wheels/source-dists ready for
upload to PyPI. A developer with the necessary credentials may then
simply download the artifact called "dist_all" and run `twine upload *`
on the unzipped contents.